Refactored image loading to accommodate for ptu files
