+++ image_handler.py
@@ -11,13 +11,15 @@ from matplotlib import cm
 from scipy import signal
 import tifffile
 
+from image_loader.image_loader import ImageLoader
+
 np.seterr(divide='ignore', invalid='ignore')
 
 class ImageHandler:
 	"""Handles the calculations for the images, taken in as tiff stacks located at filename"""
 
 	def __init__(self, filename, phi_cal=0, m_cal=1, bin_width=0.2208, freq=80, harmonic=1):
-		im = io.imread(filename)
+		im = ImageLoader.load(filename)
 		print("Shape of image: ", im.shape)
 		self.name = os.path.splitext(os.path.basename(filename))[0]
 		self.original_image = np.sum(im, axis=0)
