--- image_loader/tiff_loader.py
@@ -1,9 +1,0 @@
-from skimage import io
-
-from image_loader import ImageLoader
-
-class TiffLoader(ImageLoader):
-	
-	def load_image(self, filename):
-		im = io.imread(filename)
-		return im
